extract FileRecordReaderIterator to DataFile.scala
